feat: dto에서 domain으로 폴더 변경
